+++ utils/model_info.py
@@ -8,7 +8,7 @@ from torch.nn import parameter
 def get_model_info(model, tsize):
 
     stride = 64
-    img = torch.zeros((1, 3, stride, stride), device=next(model.parameters()).device, dtype=torch.double)
+    img = torch.zeros((1, 3, stride, stride), device=next(model.parameters()).device)
     flops, params = profile(deepcopy(model), inputs=(img,), verbose=False)
     params /= 1e6
     flops /= 1e9
